LinePlottingOptions.tsx colours select updated LineSeriesVis.tsx integrated with Options and data

+++ src/UI/LoggedIn/Plotting/Line/LinePlottingOptions.tsx
@@ -279,7 +279,7 @@ function LinePlottingOptions(props: any) {
                             onChange={(event) => {
                                 setOptions({
                                     ...options,
-                                    opacity: parseInt(event.target.value),
+                                    opacity: parseFloat(event.target.value),
                                 });
                             }}
                         />
